add: SearchTermList를 컴포넌트로 분리
