+++ client/src/components/Header/HeaderMain/SearchBar/Dropdown.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import SearchTerm from '../../../../models/searchTerm';
+import SearchTermList from './SearchTermList/SearchTermList';
 
 const Container = styled.div`
   border: 1px solid ${(props) => props.theme.color.grey1};
@@ -17,52 +18,12 @@ const DropdownHeader = styled.div`
   padding: 8px 16px 0px 16px;
 `;
 
+const DropdownBody = styled.div``;
+
 const DropdownTitle = styled.div`
   padding-bottom: 8px;
   border-bottom: 1px solid ${(props) => props.theme.color.grey1};
   background-color: ${(props) => props.theme.color.white1};
-`;
-
-const DropdownBody = styled.div`
-  padding: 8px 16px;
-  background-color: ${(props) => props.theme.color.white1};
-`;
-
-const FlexCenterWrapper = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`;
-
-const SearchTermItem = styled(FlexCenterWrapper)`
-  margin: 8px 0px;
-`;
-
-const SearchTermItemLeft = styled(FlexCenterWrapper)`
-  justify-content: flex-start;
-
-  :hover {
-    cursor: pointer;
-  }
-`;
-
-const SearchTermItemRight = styled(FlexCenterWrapper)`
-  justify-content: flex-end;
-`;
-
-const Date = styled.div`
-  font-size: 10px;
-  color: ${(propss) => propss.theme.color.grey3};
-  margin-right: 8px;
-  cursor: default;
-`;
-
-const DeleteButton = styled.div`
-  color: ${(propss) => propss.theme.color.grey5};
-
-  :hover {
-    cursor: pointer;
-  }
 `;
 
 const DropdownController = styled.div`
@@ -84,9 +45,7 @@ const DropdownButton = styled.div`
 const DropdownAllDeleteButton = styled(DropdownButton)``;
 const DropdownCloseButton = styled(DropdownButton)``;
 
-const EmptyList = styled.div`
-  margin: 8px 0px;
-`;
+const Empty = styled.div``;
 
 type Props = {
   searchTermList: SearchTerm[];
@@ -95,27 +54,8 @@ type Props = {
   onDeleteSearchTerm: (content: string) => React.MouseEventHandler;
 };
 
-const NO_SEARCHTERM_LIST_TITLE = '최근 검색어가 없습니다.';
-
 const Dropdown = (props: Props): JSX.Element => {
   const { searchTermList, onCloseDropdown, onDeleteAllSearchTerm, onDeleteSearchTerm } = props;
-
-  const SearchTermItems = searchTermList.map((searchTerm, index) => (
-    <SearchTermItem key={index}>
-      <SearchTermItemLeft data-testid={`search-term-item-content-${index}`}>
-        {searchTerm.content}
-      </SearchTermItemLeft>
-      <SearchTermItemRight>
-        <Date data-testid={`search-term-item-date-${index}`}>{searchTerm.date}</Date>
-        <DeleteButton
-          onClick={onDeleteSearchTerm(searchTerm.content)}
-          data-testid={`search-term-item-delete-btn-${index}`}
-        >
-          X
-        </DeleteButton>
-      </SearchTermItemRight>
-    </SearchTermItem>
-  ));
 
   return (
     <Container>
@@ -123,19 +63,20 @@ const Dropdown = (props: Props): JSX.Element => {
         <DropdownTitle>최근 검색어</DropdownTitle>
       </DropdownHeader>
       <DropdownBody>
-        {SearchTermItems.length === 0 ? (
-          <EmptyList>{NO_SEARCHTERM_LIST_TITLE}</EmptyList>
-        ) : (
-          SearchTermItems
-        )}
+        <SearchTermList searchTermList={searchTermList} onDeleteSearchTerm={onDeleteSearchTerm} />
       </DropdownBody>
       <DropdownController>
-        <DropdownAllDeleteButton
-          onClick={onDeleteAllSearchTerm}
-          data-testid="search-term-all-delete-btn"
-        >
-          전체 삭제
-        </DropdownAllDeleteButton>
+        {searchTermList.length === 0 ? (
+          <Empty />
+        ) : (
+          <DropdownAllDeleteButton
+            onClick={onDeleteAllSearchTerm}
+            data-testid="search-term-all-delete-btn"
+          >
+            전체 삭제
+          </DropdownAllDeleteButton>
+        )}
+
         <DropdownCloseButton onClick={onCloseDropdown} data-testid="dropdown-close-btn">
           닫기
         </DropdownCloseButton>
